// src/renderer/types.ts
export interface Basic {
  gender?: string;
  playAids?: string[];
  race?: string;
  size?: number;
  mass?: number;
  age?: number;
  eyeColor?: string;
  hairColor?: string;
  calendar?: string;
  birthDate?: string;
  startGeneratingPoints?: number;
  culture?: string;
  profession?: {
    kind?: string;
    name?: string;
    coverIdentity?: string;
    variant?: string;
  };
  notes?: string[];
  exp?: number;
  freeExp?: number;
}